# docs/openapi.yml
openapi: 3.1.0
info:
  title: Bank REST Test Assignment
  description: |-
    API specification for a bank cards web service.
    
    Features:
    - Bank cards CRUD with filters & pagination
    - Card operations (activating, blocking, transfers)
    - JWT authorization with access & refresh tokens
  version: 1.0.0
servers:
  - url: https://localhost:8443/api/v1
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registration
      description: Register a new user and get access & refresh tokens.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
      responses:
        200:
          description: User registered successfuly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        400:
          description: Password is too short or too long
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        409:
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in
      description: Log in and get access & refresh tokens.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        200:
          description: User logged in successfuly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: Incorrect email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh token
      description: Get a new access token with refresh token.
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Generated new access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        400:
          description: Bad request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: Refresh token not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out
      description: |-
        Delete user's current refresh token requiring logging in
        for future refreshes. Access token still can be used
        for operations until it is expired.
        If no refresh token is found,
        operation will still be considered successful.
      operationId: invalidateRefreshToken
      security:
        - bearerAuth: []
      responses:
        204:
          description: Successful operation
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get user info
      description: Get current user's email & role.
      operationId: selfCheck
      security:
        - bearerAuth: []
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        messages:
          type: array
          items:
              type: string
        timestamp:
          type: string
          format: date-time
    RegistrationRequest:
      type: object
      properties:
        firstName:
          type: string
          example: "Robert"
        lastName:
          type: string
          example: "Martin"
        email:
          type: string
          example: "example@gmail.com"
        password:
          type: string
          description: 8-128 characters
          example: "p4sswORD_"
      required:
        - firstName
        - lastName
        - email
        - password
    AuthRequest:
      type: object
      properties:
        email:
          type: string
          example: "example@gmail.com"
        password:
          type: string
          description: 8-128 characters
          example: "p4sswORD_"
      required:
        - email
        - password
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: UUID refresh token
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
          example: "Robert"
        lastName:
          type: string
          example: "Martin"
        email:
          type: string
          example: "example@gmail.com"
        roles:
          type: array
          items:
            type: string
            enum:
              - ROLE_USER
              - ROLE_ADMIN
            example: "ROLE_USER"
...